scripts: Restore GPG/SSH keys with restore_backup.sh

+++ scripts/restore_backup.sh
@@ -6,25 +6,37 @@ BACKUP_DIRS=(~/dotfiles/backup/$(hostname)/{pkgs,systemd_units,keys})
 echo 'Restore GPG and SSH keys?'
 select choice in "Yes" "No"; do
   case $choice in
-    Yes ) echo 'Decrypt encrypted tar archive'
-          $(cd ${BACKUP_DIRS[2]} && gpg --quiet --output keys.tar --decrypt keys.tar.gpg)
+    Yes ) DECRYPT_TRIES=3
+          TRIES=0
 
-          echo 'Reload GPG agent to clear the cached passphrase'
-          gpg-connect-agent reloadagent /bye 1> /dev/null
+          echo "Decrypting encrypted tar archive (Maximum $DECRYPT_TRIES tries)"
+          # Until X unsuccessful tries OR tar archive successfully decrypted
+          until [ $TRIES -eq $DECRYPT_TRIES ] || [ -e ${BACKUP_DIRS[2]}/keys.tar ]; do
+            gpg --quiet --output ${BACKUP_DIRS[2]}/keys.tar --decrypt ${BACKUP_DIRS[2]}/keys.tar.gpg
 
-          echo 'Extract GPG and SSH keys from unencrypted tar archive'
+            echo 'Reloading GPG agent to clear the cached passphrase'
+            gpg-connect-agent reloadagent /bye 1> /dev/null;
+
+            TRIES=$((TRIES + 1))
+          done
+
+          if [ ! -e ${BACKUP_DIRS[2]}/keys.tar ]; then
+            echo 'Failed to decrypt tar archive. Exiting...'
+            break
+          fi
+
+          echo 'Extracting GPG and SSH keys from unencrypted tar archive'
           $(cd ${BACKUP_DIRS[2]} && tar xvf keys.tar 1> /dev/null)
 
-          echo 'Restore GPG keys'
-          gpg2 --import ${BACKUP_DIRS[2]}/secrets.asc
-          # Get the key id (2E...4A from example below) and trust it
-          # gpg: key 2E558BC282FA5A4A: "Dany Marcoux <danymarcoux@gmail.com>" not changed
+          echo 'Restoring GPG keys and their trusts'
+          gpg --import ${BACKUP_DIRS[2]}/secrets.asc
+          gpg --import-ownertrust ${BACKUP_DIRS[2]}/otrust.txt
 
-          echo 'Restore SSH keys'
+          echo 'Restoring SSH keys'
           cp ${BACKUP_DIRS[2]}/*id_rsa* ~/.ssh/
 
-          echo 'Remove exported GPG/SSH keys and tar archives'
-          $(cd ${BACKUP_DIRS[2]} && rm secrets.asc *id_rsa* keys.tar{,.gpg})
+          echo 'Removing exported GPG/SSH keys and the unencrypted tar archive'
+          $(cd ${BACKUP_DIRS[2]} && rm secrets.asc otrust.txt *id_rsa* keys.tar)
 
           break ;;
     No ) break ;;
